requirements.txt
streamlit
python-dotenv
requests
opencv-python